Simplify the code
Thank you for this sample code - and providing the other information for the Advent of Code. I normally run this same thing with my students but don't open it up to the school, and don't offer any tips or sample code, etc. If you have any interest in a different style 'competition' that is less about code and more about Cyber Security, you should look at CyberPatriot next year. Would be happy to assist.

(This modification is for my grade 12 students or higher - **not** a critique of the original code. The original code is a lovely and detailed demonstration of a solution. This is an example of one method of simplification. In fact, it could be simplified further by modifying the for loop, but we'll leave that for now.)

In smaller systems (think IoT) there is no benefit to "extracting" each character into a variable and utilizing that memory space. Instead, test it directly.
For the conditional (if-statement) the ternary operator would make this cleaner. To be clear - this will still cause a "if this, else that" to occur in the background. It just **reads** cleaner in the sense that it is less lines of code. Very common in, say, JavaScript and Python.

If input validity checking is not required, you don't need to check for the ')'. If it's not "up a floor" then it's "down a floor".
The description and sample inputs from the challenge page do not explicitly state if input is always valid but the samples seem to imply that they are.

To be clear - this modification has not been run through a Java compiler, nor tested. I was bored while waiting for someone.